PlankBot-Evolved/requirements.txt
discord.py==2.0.0a3575+g45d498c1
discord==1.7.3
python-dotenv==0.19.0
